pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
dynamic = [ "version" ]
name = "azure-cost-exporter"
dependencies = [
  "azure-mgmt-costmanagement>=4.0.0,<5.0",
  "azure-identity>=1.13.0,<2.0",
  "prometheus-client>=0.17.0,<1.0",
  "envyaml==1.10.211231",
  "pyyaml>=6.0.1,<7.0.0",
  "python-dateutil>=2.8.2,<3.0",
  "coloredlogs>=15.0.0,<16.0",
]
requires-python = ">=3.9"
authors = [
  {name = "Electrolux Opensource Projects", email = "opensource@electrolux.com"},
  {name = "Victor Cabezas", email = "wiston666@gmail.com"},
]
maintainers = [
  {name = "Victor Cabezas", email = "wiston666@gmail.com"},
]
description = "Azure Costs Exporter"
readme = "README.md"
license = {file = "LICENSE"}
classifiers = [
  "Programming Language :: Python :: 3.11",
]

[project.urls]
Homepage = "https://github.com/Wiston999/azure-cost-exporter.git"
Repository = "https://github.com/Wiston999/azure-cost-exporter.git"

[project.scripts]
azure-cost-exporter = "azure_cost_exporter.cli:main"

[tool.setuptools.dynamic]
version = {attr = "azure_cost_exporter.__version__"}
